+++ src/search.h
@@ -70,7 +70,7 @@ struct SearchStack {
 extern void init_search();
 extern void init_threads();
 extern void exit_threads();
-extern int perft(Position& pos, Depth depth);
+extern int64_t perft(Position& pos, Depth depth);
 extern bool think(Position& pos, bool infinite, bool ponder, int time[], int increment[],
                   int movesToGo, int maxDepth, int maxNodes, int maxTime, Move searchMoves[]);
